Reimplemented remote action interface, added "su" and "do" commands.
Some hacks required to circumvent the terminal being left in a
confused state when a command reading input (su) is invoked from
within a signal handler:

If an action that depends on su is invoked remotely then a flag file
is used to reset the terminal on the next map by sending a newline.
See remote.run, window.map, and engine.reset.