Created plugins 'terminal' and 'dropdown', assure toplevel rc overwrites plugins.
The dump/restore order is changed so that the user configs are always
at the end of the chain, thus assuring that object properties from
plugins can be overwritten in the toplevel rcfile.

The terminal mode is left in core, along with the launcher.
(this still allows BASHRUN=1 urxvt (implies terminal) even if the
plugin is not loaded. It's mainly in a plugin so that people can
see how it works and/or change it.

Added 'terminal' property to mode. (having terminals of the same name
is pointless here)