Use commandline as the title for terminal windows
When a command is run in a terminal, the command will be used as
the window title for the terminal window.

Added --title option for terminal objects. If specified, the given
title will be used instead of the actual commandline.

Added "title" command for action functions to specify the title for
the terminal window directly. This is useful if the actual command
line differs from the entered command line.

This feature enhances working with a lot of terminal windows running
console applications, which now have more meaningful titles instead of
all being titled "bash". Setting the title also allows using different
icons for pagers or taskbars. For example, when using urxvt, the icon
can be configured individually in .Xdefaults:

urxvt*iconFile: /path/to/terminal.png
htop*iconFile: /path/to/htop.png
emacs*iconFile: /path/to/emacs.png
alsamixer*iconFile: /path/to/alsamixer.png