Added support for automatically loading icons from icon directories
On startup, if either $HOME/.config/bashrun2/icons or
$SYSCONFDIR/xdg/bashrun2/icons exists, bashrun will add an X resource
for each .png file found in these directories, in the format

<basename-of-icon>*iconFile: <full-path-of-icon>

Also, every time a command is launched in the terminal, bashrun will
attempt to load an icon resource in the same manner, based on the
first word on the commandline.

Icons from the users configuration directory take precedence over
those found in the system-wide configuration directory.